edited staffmember routes to plural. GET requests for single and all staffmembers.
